+++ routes/web.php
@@ -1,10 +1,10 @@
 <?php
 
-use App\Http\Controllers\Authentication;
-use App\Http\Controllers\Game;
-use App\Http\Controllers\Index;
-use App\Http\Controllers\Player;
-use App\Http\Controllers\Share;
+use App\Http\Controllers\View\Authentication;
+use App\Http\Controllers\View\Game;
+use App\Http\Controllers\View\Index;
+use App\Http\Controllers\View\Player;
+use App\Http\Controllers\View\Share;
 use Illuminate\Support\Facades\Route;
 
 Route::get(
@@ -14,8 +14,8 @@ Route::get(
 
 Route::post(
     '/create-password',
-    [Authentication::class, 'createPasswordProcess']
-)->name('create-password.process');
+    [\App\Http\Controllers\Action\Authentication::class, 'createPassword']
+)->name('create-password.action');
 
 Route::get(
     '/',
@@ -29,8 +29,8 @@ Route::get(
 
 Route::post(
     '/sign-in',
-    [Authentication::class, 'signInProcess']
-)->name('sign-in.process');
+    [\App\Http\Controllers\Action\Authentication::class, 'signIn']
+)->name('sign-in.action');
 
 Route::get(
     '/register',
@@ -39,8 +39,8 @@ Route::get(
 
 Route::post(
     '/register',
-    [Authentication::class, 'registerProcess']
-)->name('register.process');
+    [\App\Http\Controllers\Action\Authentication::class, 'register']
+)->name('register.action');
 
 Route::get(
     '/registration-complete',
@@ -59,13 +59,13 @@ Route::get(
 
 Route::post(
     '/public/score-sheet/{token}/score-upper',
-    [Share::class, 'scoreUpper']
-)->name('public.score-upper');
+    [\App\Http\Controllers\Action\Share::class, 'scoreUpper']
+)->name('public.score-upper.action');
 
 Route::post(
     '/public/score-sheet/{token}/score-lower',
-    [Share::class, 'scoreLower']
-)->name('public.score-lower');
+    [\App\Http\Controllers\Action\Share::class, 'scoreLower']
+)->name('public.score-lower.action');
 
 Route::get(
     '/public/game/{token}/player-scores',
@@ -89,6 +89,10 @@ Route::group(
             [Index::class, 'home']
         )->name('home');
 
+        Route::post(
+            '/start',
+            [\App\Http\Controllers\Action\Game::class, 'start']
+        )->name('start');
 
         Route::get(
             '/new-game',
@@ -97,8 +101,8 @@ Route::group(
 
         Route::post(
             '/new-game',
-            [Game::class, 'newGameProcess']
-        )->name('game.create.process');
+            [\App\Http\Controllers\Action\Game::class, 'newGame']
+        )->name('game.create.action');
 
         Route::get(
             '/game/{game_id}/player/{player_id}/score-sheet',
@@ -117,23 +121,23 @@ Route::group(
 
         Route::post(
             '/game/{game_id}/complete',
-            [Game::class, 'complete']
-        )->name('game.complete');
+            [\App\Http\Controllers\Action\Game::class, 'complete']
+        )->name('game.complete.action');
 
         Route::post(
             '/game/{game_id}/complete-and-play-again',
-            [Game::class, 'completeAndPlayAgain']
-        )->name('game.complete.play-again');
+            [\App\Http\Controllers\Action\Game::class, 'completeAndPlayAgain']
+        )->name('game.complete.play-again.action');
 
         Route::post(
             '/game/{game_id}/delete',
-            [Game::class, 'deleteGame']
-        )->name('game.delete');
+            [\App\Http\Controllers\Action\Game::class, 'deleteGame']
+        )->name('game.delete.action');
 
         Route::post(
             '/game/score-upper',
-            [Game::class, 'scoreUpper']
-        )->name('game.score-upper');
+            [\App\Http\Controllers\Action\Game::class, 'scoreUpper']
+        )->name('game.score-upper.action');
 
         Route::get(
             '/game/{game_id}/player/{player_id}/bonus',
@@ -142,13 +146,13 @@ Route::group(
 
         Route::get(
             '/game/{game_id}/player/{player_id}/delete',
-            [Game::class, 'deleteGamePlayer']
+            [\App\Http\Controllers\Action\Game::class, 'deleteGamePlayer']
         )->name('game.player.delete');
 
         Route::post(
             '/game/score-lower',
-            [Game::class, 'scoreLower']
-        )->name('game.score-lower');
+            [\App\Http\Controllers\Action\Game::class, 'scoreLower']
+        )->name('game.score-lower.action');
 
         Route::get(
             '/game/{game_id}/player-scores',
@@ -162,8 +166,8 @@ Route::group(
 
         Route::post(
             '/add-players-to-game/{game_id}',
-            [Game::class, 'addPlayersToGameProcess']
-        )->name('game.add-players.process');
+            [\App\Http\Controllers\Action\Game::class, 'addPlayersToGame']
+        )->name('game.add-players.action');
 
 
         Route::get(
@@ -178,8 +182,8 @@ Route::group(
 
         Route::post(
             '/new-player',
-            [Player::class, 'newPlayerProcess']
-        )->name('player.create.process');
+            [\App\Http\Controllers\Action\Player::class, 'newPlayer']
+        )->name('player.create.action');
 
         Route::get(
             '/account',
@@ -193,12 +197,17 @@ Route::group(
 
         Route::post(
             '/account/delete-yatzy-account',
-            [Authentication::class, 'deleteYatzyAccount']
-        )->name('account.delete-yatzy-account');
+            [\App\Http\Controllers\Action\Authentication::class, 'deleteYatzyAccount']
+        )->name('account.delete-yatzy-account.action');
 
         Route::get(
             '/account/confirm-delete-account',
             [Authentication::class, 'confirmDeleteAccount']
         )->name('account.confirm-delete-account');
+
+        Route::post(
+            '/account/delete-account',
+            [\App\Http\Controllers\Action\Authentication::class, 'deleteAccount']
+        )->name('account.delete-account.action');
     }
 );